refractor: split to ES2015 modules
